packages/sdk-python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tracion"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.24.0"]
openai = ["openai>=1.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
